Removed types dir and made types global in d.ts file
