--- src/components/types/SkcApiTypes.d.ts
@@ -1,20 +1,0 @@
-declare type ProductInfo = {
-	productId: string,
-	productLocale?: string,
-	productName?: string,
-	productType?: string,
-	productSubType?: string,
-	productReleaseDate: string,
-	productTotal?: number,
-	productContent: ProductContent[]
-}
-
-declare type ProductContent = {
-	productPosition: string,
-	rarities: string[]
-	card?: Card
-}
-
-declare type Card = {
-
-}
